Improve finding half.hpp
On Fedora half.hpp is installed to the system dir /usr/include.
There is no cmake package, so finding the half cmake package will always fail.
The source looks for half with #include <half/half.h>  The use of half/ makes
it impossible to find and use the system location.

So look for the half.hpp path and add add that to the include_directories.

Use this sed script to programatically change the uses of #include <half/half.h>

  for f in `find . -type f -name '*.hpp' -o -name '*.cpp' `; do
      sed -i -e 's@#include <half/half.hpp>@#include <half.hpp>@' $f
  done

Signed-off-by: Tom Rix <trix@redhat.com>