Improve compiler setup
If the user has passed in the compiler to use for building, use it.
Otherwise set it to clang++

On Fedora there is no /opt/rocm/llvm/bin/clang++, the compiler used is
a system compiler wrapped by hipcc. So set COMPILER_FOR_HIP to llvm/bin/clang++
only if that file exists, otherwise use the CMAKE_CXX_COMPILER.

Signed-off-by: Tom Rix <trix@redhat.com>